+++ app/src/main/java/com/cosmos/unreddit/ui/mediaviewer/MediaViewerFragment.kt
@@ -251,6 +251,8 @@ class MediaViewerFragment : FullscreenBottomSheetFragment() {
 
         videoViewHolder?.muteAudio(shouldMute)
 
+        mediaAdapter.muteVideo = shouldMute
+
         viewerViewModel.setMuted(shouldMute)
     }