+++ buildSrc/src/main/kotlin/Config.kt
@@ -1,0 +1,12 @@
+object Config {
+    const val namespace = "com.cosmos.unreddit"
+
+    const val applicationId = "com.cosmos.unreddit"
+
+    const val minSdk = 21
+    const val compileSdk = 33
+    const val targetSdk = 33
+
+    const val versionCode = 17
+    const val versionName = "2.1.0"
+}